requirements.txt
matplotlib>=3.7.0
pandas>=2.0.0
requests>=2.31.0
google-genai
plotly>=5.24.1
kaleido==0.2.1
